Move the command-line code to its own file
