+++ pyproject.toml
@@ -25,12 +25,14 @@ classifiers = [
     "Programming Language :: Python :: 3.9",
     "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
+    "Topic :: Software Development :: Libraries",
+    "Topic :: Software Development :: Libraries :: Python Modules",
     "Topic :: Utilities",
 ]
 dynamic = ["version"]
 
 [project.scripts]
-pakler = "pakler.__init__:main"
+pakler = "pakler.__main__:main"
 
 [project.urls]
 Homepage = "https://github.com/vmallet/pakler"
